Support UnixDomainSocketEndPoint on Windows
Windows is adding support for Unix domain sockets, and it's easy to enable that support via our new UnixDomainSocketEndPoint type.
