+++ src/System.Net.Sockets/src/System/Net/Sockets/UnixDomainSocketEndPoint.Unix.cs
@@ -8,17 +8,11 @@ using System.Text;
 namespace System.Net.Sockets
 {
     /// <summary>Represents a Unix Domain Socket endpoint as a path.</summary>
-    public sealed class UnixDomainSocketEndPoint : EndPoint
+    public sealed partial class UnixDomainSocketEndPoint : EndPoint
     {
-        private const AddressFamily EndPointAddressFamily = AddressFamily.Unix;
-
-        private static readonly Encoding s_pathEncoding = Encoding.UTF8;
         private static readonly int s_nativePathOffset;
         private static readonly int s_nativePathLength;
         private static readonly int s_nativeAddressSize;
-
-        private readonly string _path;
-        private readonly byte[] _encodedPath;
 
         static UnixDomainSocketEndPoint()
         {
@@ -28,108 +22,5 @@ namespace System.Net.Sockets
             Debug.Assert(s_nativePathOffset + s_nativePathLength <= s_nativeAddressSize, "Expected address size to include all of the path length");
             Debug.Assert(s_nativePathLength >= 92, "Expected max path length to be at least 92"); // per http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/sys_un.h.html
         }
-
-        public UnixDomainSocketEndPoint(string path)
-        {
-            if (path == null)
-            {
-                throw new ArgumentNullException(nameof(path));
-            }
-
-            // Pathname socket addresses should be null-terminated.
-            // Linux abstract socket addresses start with a zero byte, they must not be null-terminated.
-            bool isAbstract = IsAbstract(path);
-            int bufferLength = s_pathEncoding.GetByteCount(path);
-            if (!isAbstract)
-            {
-                // for null terminator
-                bufferLength++;
-            }
-
-            if (path.Length == 0 || bufferLength > s_nativePathLength)
-            {
-                throw new ArgumentOutOfRangeException(
-                    nameof(path), path, 
-                    SR.Format(SR.ArgumentOutOfRange_PathLengthInvalid, path, s_nativePathLength));
-            }
-
-            _path = path;
-            _encodedPath = new byte[bufferLength];
-            int bytesEncoded = s_pathEncoding.GetBytes(path, 0, path.Length, _encodedPath, 0);
-            Debug.Assert(bufferLength - (isAbstract ? 0 : 1) == bytesEncoded);
-        }
-
-        internal UnixDomainSocketEndPoint(SocketAddress socketAddress)
-        {
-            if (socketAddress == null)
-            {
-                throw new ArgumentNullException(nameof(socketAddress));
-            }
-
-            if (socketAddress.Family != EndPointAddressFamily || 
-                socketAddress.Size > s_nativeAddressSize)
-            {
-                throw new ArgumentOutOfRangeException(nameof(socketAddress));
-            }
-
-            if (socketAddress.Size > s_nativePathOffset)
-            {
-                _encodedPath = new byte[socketAddress.Size - s_nativePathOffset];
-                for (int i = 0; i < _encodedPath.Length; i++)
-                {
-                    _encodedPath[i] = socketAddress[s_nativePathOffset + i];
-                }
-
-                // Strip trailing null of pathname socket addresses.
-                int length = _encodedPath.Length;
-                if (!IsAbstract(_encodedPath))
-                {
-                    // Since this isn't an abstract path, we're sure our first byte isn't 0.
-                    while (_encodedPath[length - 1] == 0)
-                    {
-                        length--;
-                    }
-                }
-                _path = s_pathEncoding.GetString(_encodedPath, 0, length);
-            }
-            else
-            {
-                _encodedPath = Array.Empty<byte>();
-                _path = string.Empty;
-            }
-        }
-
-        public override SocketAddress Serialize()
-        {
-            var result = new SocketAddress(AddressFamily.Unix, s_nativePathOffset + _encodedPath.Length);
-
-            for (int index = 0; index < _encodedPath.Length; index++)
-            {
-                result[s_nativePathOffset + index] = _encodedPath[index];
-            }
-
-            return result;
-        }
-
-        public override EndPoint Create(SocketAddress socketAddress) => new UnixDomainSocketEndPoint(socketAddress);
-
-        public override AddressFamily AddressFamily => EndPointAddressFamily;
-
-        public override string ToString()
-        {
-            bool isAbstract = IsAbstract(_path);
-            if (isAbstract)
-            {
-                return "@" + _path.Substring(1);
-            }
-            else
-            {
-                return _path;
-            }
-        }
-
-        private static bool IsAbstract(string path) => path.Length > 0 && path[0] == '\0';
-
-        private static bool IsAbstract(byte[] encodedPath) => encodedPath.Length > 0 && encodedPath[0] == 0;
     }
 }
