+++ src/System.Net.Sockets/src/System/Net/Sockets/UnixDomainSocketEndPoint.Windows.cs
@@ -5,11 +5,19 @@
 namespace System.Net.Sockets
 {
     /// <summary>Represents a Unix Domain Socket endpoint as a path.</summary>
-    public sealed class UnixDomainSocketEndPoint : EndPoint
+    public sealed partial class UnixDomainSocketEndPoint : EndPoint
     {
-        public UnixDomainSocketEndPoint(string path)
-        {
-            throw new PlatformNotSupportedException();
-        }
+        private static readonly int s_nativePathOffset = 2; // sizeof(sun_family)
+        private static readonly int s_nativePathLength = 108; // sizeof(sun_path)
+        private static readonly int s_nativeAddressSize = s_nativePathOffset + s_nativePathLength; // sizeof(sockaddr_un)
+
+        // from afunix.h:
+        //#define UNIX_PATH_MAX 108
+        //typedef struct sockaddr_un
+        //{
+        //    ADDRESS_FAMILY sun_family;     /* AF_UNIX */
+        //    char sun_path[UNIX_PATH_MAX];  /* pathname */
+        //}
+        //SOCKADDR_UN, *PSOCKADDR_UN;
     }
 }
